Add read/write ECR policy
This creates a single read/write policy rather than having to use a
combination of three different policies. In most cases, this is the
policy we would use, so this will make using it easier.

I have also refined the list of actions based on the documentation to
reflect what seems like the appropriate and necessary permissions.